Refs #29940 - Refactor select options
